[linux] network: move implementation to platform folder.
