+++ cmake/treedata/android/subdirs.txt
@@ -2,7 +2,6 @@ xbmc/cores/RetroPlayer/process/android  cores/RetroPlayer/process/android
 xbmc/platform/linux                     platform/linux
 xbmc/input/touch                        input/touch
 xbmc/input/touch/generic                input/touch/generic
-xbmc/network/linux                      network/linux
 xbmc/peripherals/bus/linux              peripherals/bus/linux
 xbmc/peripherals/bus/android            peripherals/bus/android
 xbmc/powermanagement/android            powermanagement/android
@@ -10,6 +9,7 @@ xbmc/filesystem/posix                   filesystem/posix
 xbmc/utils/posix                        utils_posix
 xbmc/windowing/android                  windowing/android
 xbmc/platform/posix                     posix
+xbmc/platform/linux/network             platform/linux/network
 xbmc/platform/android/activity          android_activity
 xbmc/platform/android/bionic_supplement android_bionicsupplement
 xbmc/platform/android/network           platform/android/network
