+++ cmake/treedata/ios/subdirs.txt
@@ -1,7 +1,7 @@
 xbmc/platform/linux              platform/linux
+xbmc/platform/linux/network      platform/linux/network
 xbmc/input/touch                 input/touch
 xbmc/input/touch/generic         input/touch/generic
-xbmc/network/linux               network/linux
 xbmc/network/osx                 network/osx
 xbmc/peripherals/bus/osx         peripherals/bus/osx
 xbmc/powermanagement/osx         powermanagement/osx
