+++ xbmc/platform/linux/network/NetworkLinux.h
