+++ xbmc/platform/linux/network/ZeroconfAvahi.h
