+++ xbmc/platform/linux/network/ZeroconfBrowserAvahi.h
